+++ fastapi_mvc/generators/project/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/__main__.py
@@ -1,6 +1,0 @@
-"""{{cookiecutter.project_description}}"""
-from {{cookiecutter.package_name}}.cli import cli
-
-
-if __name__ == "__main__":
-    cli()