+++ fastapi_mvc/generators/project/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/app/views/__init__.py
@@ -1,6 +1,0 @@
-"""Application implementation - views."""
-from {{cookiecutter.package_name}}.app.views.error import ErrorResponse
-from {{cookiecutter.package_name}}.app.views.ready import ReadyResponse
-
-
-__all__ = ("ErrorResponse", "ReadyResponse")